refactor tools back to sensor / eeffector
